+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/RoomHeader.kt
@@ -2,6 +2,7 @@ package de.connect2x.messenger.compose.view.room.timeline
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
@@ -10,6 +11,7 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentSize
@@ -34,6 +36,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
@@ -50,9 +55,10 @@ import de.connect2x.messenger.compose.view.common.icons.UnencryptedIcon
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.isMobile
+import de.connect2x.messenger.compose.view.root.IsSinglePane
+import de.connect2x.messenger.compose.view.theme.MaxHeaderHeight
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.RoomHeaderInfo
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.RoomHeaderViewModel
-
 
 interface RoomHeaderView {
     @Composable
@@ -78,12 +84,18 @@ class RoomHeaderViewImpl : RoomHeaderView {
         val roomHeaderElement = roomHeaderViewModel.roomHeaderInfo.collectAsState().value
         val usersTyping = roomHeaderViewModel.usersTyping.collectAsState().value
         val isDirectChat = roomHeaderViewModel.isDirectChat.collectAsState().value
+        val headerHeightFlow = MaxHeaderHeight.current
+        val headerHeight = headerHeightFlow.collectAsState().value
+        val density = LocalDensity.current
+
         Surface(color = MaterialTheme.colorScheme.surface, tonalElevation = 8.dp) {
             Column {
                 Row(
-                    Modifier
-                        .fillMaxWidth()
-                        .align(Alignment.CenterHorizontally)
+                    horizontalArrangement = Arrangement.Center,
+                    modifier = Modifier.fillMaxWidth().onGloballyPositioned { coordinates ->
+                        val newHeaderHeight = with(density) { coordinates.size.height.toDp() - 1.toDp() }
+                        headerHeightFlow.value = maxOf(headerHeight, newHeaderHeight)
+                    }
                 ) {
                     if (showBackButton) {
                         RoomBackButton(roomHeaderViewModel)
@@ -130,6 +142,18 @@ class RoomHeaderViewImpl : RoomHeaderView {
                             }
                         }
                         RoomExtras(roomHeaderViewModel, showSettingsButton)
+                    }
+
+                    // If we have a multi-pane view, we will display an invisible text that has the function of
+                    // forcing the three header elements to the same height.
+                    val density = LocalDensity.current
+                    if (!IsSinglePane.current) {
+                        Text(
+                            text = " ",
+                            style = MaterialTheme.typography.labelMedium
+                                .copy(color = MaterialTheme.colorScheme.onBackground),
+                            modifier = Modifier.height(headerHeight)
+                        )
                     }
                 }
                 HorizontalDivider(Modifier.fillMaxWidth())
@@ -186,8 +210,7 @@ fun ColumnScope.RoomTopic(roomHeaderElement: RoomHeaderInfo) {
     }) {
         Text(
             topic,
-            style = MaterialTheme.typography.titleSmall,
-            fontWeight = FontWeight.Light,
+            style = MaterialTheme.typography.labelMedium.copy(color = MaterialTheme.colorScheme.onBackground),
             overflow = TextOverflow.Ellipsis,
             maxLines = 1,
         )