+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/root/MainView.kt
@@ -3,8 +3,9 @@ package de.connect2x.messenger.compose.view.root
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
@@ -12,8 +13,7 @@ import de.connect2x.messenger.compose.view.SINGLE_PANE_THRESHOLD
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.trixnity.messenger.viewmodel.MainViewModel
 
-
-private val MAX_WIDTH = 1600.dp
+val IsSinglePane = compositionLocalOf<Boolean> { error("compositionLocal not defined") }
 
 interface MainView {
     @Composable
@@ -28,16 +28,12 @@ fun Main(mainViewModel: MainViewModel) {
 class MainViewImpl : MainView {
     @Composable
     override fun create(mainViewModel: MainViewModel) {
-        BoxWithConstraints(
-            Modifier.fillMaxSize()
-        ) {
-            Box(
-                Modifier.fillMaxSize()
-                    // TODO: remove the padding and adjust the bubble layout as it makes things look broken.
-                    .padding(horizontal = if (maxWidth - MAX_WIDTH > 0.dp) (maxWidth - MAX_WIDTH) / 2 else 0.dp)
-            ) {
+        BoxWithConstraints(Modifier.fillMaxSize()) {
+            Box(Modifier.fillMaxSize()) {
                 val isSinglePane = this@BoxWithConstraints.maxWidth < SINGLE_PANE_THRESHOLD.dp
-                InitialSyncSwitch(mainViewModel, isSinglePane)
+                CompositionLocalProvider(IsSinglePane provides isSinglePane) {
+                    InitialSyncSwitch(mainViewModel, isSinglePane)
+                }
             }
         }
         DeviceVerificationSwitch(mainViewModel)