+++ CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Font and control size settings
+
 ### Changed
 
+- Insert formula in padding of the timeline
 - Increase size of avatar and change order of settings in profile view
 - Make room notification settings collapsable and change setting design to the one used by other room settings